+++ src/parse.rs
@@ -2,12 +2,13 @@ use alloc::string::{String, ToString};
 use alloc::vec::Vec;
 use core::ops::Range;
 
-use xmlparser::{self, Reference, StrSpan, Stream, TextPos};
-
 use crate::{
     AttributeData, Document, ExpandedNameIndexed, NamespaceIdx, Namespaces, NodeData, NodeId,
-    NodeKind, ShortRange, StringStorage, NS_XMLNS_URI, NS_XML_PREFIX, NS_XML_URI, PI, XMLNS,
+    NodeKind, ShortRange, StringStorage, TextPos, NS_XMLNS_URI, NS_XML_PREFIX, NS_XML_URI, PI,
+    XMLNS,
 };
+
+use crate::tokenizer::{self, Reference, StrSpan, Stream};
 
 type Result<T> = core::result::Result<T, Error>;
 
@@ -37,12 +38,10 @@ pub enum Error {
     UnknownNamespace(String, TextPos),
 
     /// Incorrect tree structure.
+    ///
+    /// expected, actual, position
     #[allow(missing_docs)]
-    UnexpectedCloseTag {
-        expected: String,
-        actual: String,
-        pos: TextPos,
-    },
+    UnexpectedCloseTag(String, String, TextPos),
 
     /// Entity value starts with a close tag.
     ///
@@ -99,8 +98,48 @@ pub enum Error {
     /// Indicates that too many namespaces were parsed.
     NamespacesLimitReached,
 
-    /// Errors detected by the `xmlparser` crate.
-    ParserError(xmlparser::Error),
+    /// The steam ended earlier than we expected.
+    ///
+    /// Should only appear on invalid input data.
+    /// Errors in a valid XML should be handled by errors below.
+    UnexpectedEndOfStream,
+
+    /// An invalid name.
+    InvalidName(TextPos),
+
+    /// A non-XML character has occurred.
+    ///
+    /// Valid characters are: <https://www.w3.org/TR/xml/#char32>
+    NonXmlChar(char, TextPos),
+
+    /// An invalid/unexpected character.
+    ///
+    /// expected, actual, position
+    InvalidChar(u8, u8, TextPos),
+
+    /// An invalid/unexpected character.
+    ///
+    /// expected, actual, position
+    InvalidChar2(&'static str, u8, TextPos),
+
+    /// An unexpected string.
+    ///
+    /// Contains what string was expected.
+    InvalidString(&'static str, TextPos),
+
+    /// An invalid ExternalID in the DTD.
+    InvalidExternalID(TextPos),
+
+    /// A comment cannot contain `--` or end with `-`.
+    InvalidComment(TextPos),
+
+    /// A Character Data node contains an invalid data.
+    ///
+    /// Currently, only `]]>` is not allowed.
+    InvalidCharacterData(TextPos),
+
+    /// An unknown token.
+    UnknownToken(TextPos),
 }
 
 impl Error {
@@ -111,24 +150,32 @@ impl Error {
             Error::UnexpectedXmlUri(pos) => pos,
             Error::UnexpectedXmlnsUri(pos) => pos,
             Error::InvalidElementNamePrefix(pos) => pos,
-            Error::DuplicatedNamespace(ref _name, pos) => pos,
-            Error::UnknownNamespace(ref _name, pos) => pos,
-            Error::UnexpectedCloseTag { pos, .. } => pos,
+            Error::DuplicatedNamespace(_, pos) => pos,
+            Error::UnknownNamespace(_, pos) => pos,
+            Error::UnexpectedCloseTag(_, _, pos) => pos,
             Error::UnexpectedEntityCloseTag(pos) => pos,
-            Error::UnknownEntityReference(ref _name, pos) => pos,
+            Error::UnknownEntityReference(_, pos) => pos,
             Error::MalformedEntityReference(pos) => pos,
             Error::EntityReferenceLoop(pos) => pos,
             Error::InvalidAttributeValue(pos) => pos,
-            Error::DuplicatedAttribute(ref _name, pos) => pos,
-            Error::ParserError(ref err) => err.pos(),
-            _ => TextPos::new(1, 1),
+            Error::DuplicatedAttribute(_, pos) => pos,
+            Error::NoRootNode => TextPos::new(1, 1),
+            Error::UnclosedRootNode => TextPos::new(1, 1),
+            Error::DtdDetected => TextPos::new(1, 1),
+            Error::NodesLimitReached => TextPos::new(1, 1),
+            Error::AttributesLimitReached => TextPos::new(1, 1),
+            Error::NamespacesLimitReached => TextPos::new(1, 1),
+            Error::UnexpectedEndOfStream => TextPos::new(1, 1),
+            Error::InvalidName(pos) => pos,
+            Error::NonXmlChar(_, pos) => pos,
+            Error::InvalidChar(_, _, pos) => pos,
+            Error::InvalidChar2(_, _, pos) => pos,
+            Error::InvalidString(_, pos) => pos,
+            Error::InvalidExternalID(pos) => pos,
+            Error::InvalidComment(pos) => pos,
+            Error::InvalidCharacterData(pos) => pos,
+            Error::UnknownToken(pos) => pos,
         }
-    }
-}
-
-impl From<xmlparser::Error> for Error {
-    fn from(e: xmlparser::Error) -> Self {
-        Error::ParserError(e)
     }
 }
 
@@ -165,11 +212,7 @@ impl core::fmt::Display for Error {
             Error::UnknownNamespace(ref name, pos) => {
                 write!(f, "an unknown namespace prefix '{}' at {}", name, pos)
             }
-            Error::UnexpectedCloseTag {
-                ref expected,
-                ref actual,
-                pos,
-            } => {
+            Error::UnexpectedCloseTag(ref expected, ref actual, pos) => {
                 write!(
                     f,
                     "expected '{}' tag, not '{}' at {}",
@@ -212,8 +255,43 @@ impl core::fmt::Display for Error {
             Error::NamespacesLimitReached => {
                 write!(f, "more than 2^16 unique namespaces were parsed")
             }
-            Error::ParserError(ref err) => {
-                write!(f, "{}", err)
+            Error::UnexpectedEndOfStream => {
+                write!(f, "unexpected end of stream")
+            }
+            Error::InvalidName(pos) => {
+                write!(f, "invalid name token at {}", pos)
+            }
+            Error::NonXmlChar(c, pos) => {
+                write!(f, "a non-XML character {:?} found at {}", c, pos)
+            }
+            Error::InvalidChar(expected, actual, pos) => {
+                write!(
+                    f,
+                    "expected '{}' not '{}' at {}",
+                    expected as char, actual as char, pos
+                )
+            }
+            Error::InvalidChar2(expected, actual, pos) => {
+                write!(
+                    f,
+                    "expected {} not '{}' at {}",
+                    expected, actual as char, pos
+                )
+            }
+            Error::InvalidString(expected, pos) => {
+                write!(f, "expected '{}' at {}", expected, pos)
+            }
+            Error::InvalidExternalID(pos) => {
+                write!(f, "invalid ExternalID at {}", pos)
+            }
+            Error::InvalidComment(pos) => {
+                write!(f, "comment at {} contains '--'", pos)
+            }
+            Error::InvalidCharacterData(pos) => {
+                write!(f, "']]>' at {} is not allowed inside a character data", pos)
+            }
+            Error::UnknownToken(pos) => {
+                write!(f, "unknown token at {}", pos)
             }
         }
     }
@@ -264,8 +342,8 @@ impl Default for ParsingOptions {
 }
 
 struct TempAttributeData<'input> {
-    prefix: StrSpan<'input>,
-    local: StrSpan<'input>,
+    prefix: &'input str,
+    local: &'input str,
     value: StringStorage<'input>,
     #[cfg(feature = "positions")]
     pos: usize,
@@ -315,24 +393,21 @@ struct Entity<'input> {
 
 #[derive(Clone, Copy)]
 struct TagNameSpan<'input> {
-    prefix: StrSpan<'input>,
-    name: StrSpan<'input>,
-    span: StrSpan<'input>,
+    prefix: &'input str,
+    name: &'input str,
+    pos: usize,
+    prefix_pos: usize,
 }
 
 impl<'input> TagNameSpan<'input> {
     #[inline]
     fn new_null() -> Self {
         Self {
-            prefix: StrSpan::from(""),
-            name: StrSpan::from(""),
-            span: StrSpan::from(""),
+            prefix: "",
+            name: "",
+            pos: 0,
+            prefix_pos: 0,
         }
-    }
-
-    #[inline]
-    fn new(prefix: StrSpan<'input>, name: StrSpan<'input>, span: StrSpan<'input>) -> Self {
-        Self { prefix, name, span }
     }
 }
 
@@ -465,13 +540,8 @@ impl<'input> Context<'input> {
         Ok(new_child_id)
     }
 
-    fn err_pos_from_span(&self, span: StrSpan) -> TextPos {
-        Stream::from_substr(self.doc.text, span.range()).gen_text_pos()
-    }
-
-    fn err_pos_from_qname(&self, prefix: StrSpan, local: StrSpan) -> TextPos {
-        let err_span = if prefix.is_empty() { local } else { prefix };
-        self.err_pos_from_span(err_span)
+    fn err_pos_at(&self, pos: usize) -> TextPos {
+        self.doc.text_pos_at(pos)
     }
 }
 
@@ -515,10 +585,9 @@ fn parse(text: &str, opt: ParsingOptions) -> Result<Document> {
         loop_detector: LoopDetector::default(),
         doc,
     };
-
-    let parser = xmlparser::Tokenizer::from(text);
     ctx.parent_prefixes.push("");
-    process_tokens(parser, &mut ctx)?;
+
+    tokenizer::parse(text, opt.allow_dtd, &mut ctx)?;
 
     let mut doc = ctx.doc;
     if !doc.root().children().any(|n| n.is_element()) {
@@ -536,111 +605,72 @@ fn parse(text: &str, opt: ParsingOptions) -> Result<Document> {
     Ok(doc)
 }
 
-#[allow(clippy::collapsible_match)]
-fn process_tokens<'input>(
-    parser: xmlparser::Tokenizer<'input>,
-    ctx: &mut Context<'input>,
-) -> Result<()> {
-    for token in parser {
-        let token = token?;
+impl<'input> tokenizer::XmlEvents<'input> for Context<'input> {
+    fn token(&mut self, token: tokenizer::Token<'input>) -> Result<()> {
         match token {
-            xmlparser::Token::ProcessingInstruction {
-                target,
-                content,
-                span,
-            } => {
-                let pi = NodeKind::PI(PI {
-                    target: target.as_str(),
-                    value: content.map(|v| v.as_str()),
+            tokenizer::Token::ProcessingInstruction(target, value, range) => {
+                let pi = NodeKind::PI(PI { target, value });
+                self.append_node(pi, range)?;
+                self.after_text = false;
+            }
+            tokenizer::Token::Comment(text, range) => {
+                self.append_node(NodeKind::Comment(StringStorage::Borrowed(text)), range)?;
+                self.after_text = false;
+            }
+            tokenizer::Token::EntityDeclaration(name, definition) => {
+                self.entities.push(Entity {
+                    name,
+                    value: definition,
                 });
-                ctx.append_node(pi, span.range())?;
             }
-            xmlparser::Token::Comment { text, span } => {
-                ctx.append_node(
-                    NodeKind::Comment(StringStorage::Borrowed(text.as_str())),
-                    span.range(),
-                )?;
-            }
-            xmlparser::Token::Text { text } => {
-                process_text(text, ctx)?;
-            }
-            xmlparser::Token::Cdata { text, span } => {
-                process_cdata(text, span, ctx)?;
-            }
-            xmlparser::Token::ElementStart {
-                prefix,
-                local,
-                span,
-            } => {
-                if prefix.as_str() == XMLNS {
-                    let pos = ctx.err_pos_from_span(prefix);
+            tokenizer::Token::ElementStart(prefix, local, start) => {
+                if prefix == XMLNS {
+                    let pos = self.err_pos_at(start + 1);
                     return Err(Error::InvalidElementNamePrefix(pos));
                 }
 
-                ctx.tag_name = TagNameSpan::new(prefix, local, span);
+                self.tag_name = TagNameSpan {
+                    prefix,
+                    name: local,
+                    pos: start,
+                    prefix_pos: start + 1,
+                };
+
+                self.after_text = false;
             }
-            xmlparser::Token::Attribute {
-                prefix,
-                local,
-                value,
-                span,
-            } => {
-                process_attribute(prefix, local, value, span, ctx)?;
+            tokenizer::Token::Attribute(attr_start, prefix, local, value) => {
+                process_attribute(attr_start, prefix, local, value, self)?;
             }
-            xmlparser::Token::ElementEnd { end, span } => {
-                process_element(end, span, ctx)?;
+            tokenizer::Token::ElementEnd(end, range) => {
+                process_element(end, range, self)?;
+                self.after_text = false;
             }
-            xmlparser::Token::DtdStart { .. } => {
-                if !ctx.opt.allow_dtd {
-                    return Err(Error::DtdDetected);
-                }
+            tokenizer::Token::Text(text, range) => {
+                process_text(text, range, self)?;
             }
-            xmlparser::Token::EntityDeclaration {
-                name, definition, ..
-            } => {
-                if let xmlparser::EntityDefinition::EntityValue(value) = definition {
-                    ctx.entities.push(Entity {
-                        name: name.as_str(),
-                        value,
-                    });
-                }
+            tokenizer::Token::Cdata(text, range) => {
+                process_cdata(text, range, self)?;
             }
-            _ => {}
         }
 
-        match token {
-            xmlparser::Token::ProcessingInstruction { .. }
-            | xmlparser::Token::Comment { .. }
-            | xmlparser::Token::ElementStart { .. }
-            | xmlparser::Token::ElementEnd { .. } => {
-                ctx.after_text = false;
-            }
-            _ => {}
-        }
+        Ok(())
     }
-
-    Ok(())
 }
 
 #[allow(clippy::too_many_arguments)]
 fn process_attribute<'input>(
-    prefix: StrSpan<'input>,
-    local: StrSpan<'input>,
+    attr_pos: usize,
+    prefix: &'input str,
+    local: &'input str,
     value: StrSpan<'input>,
-    token_span: StrSpan<'input>,
     ctx: &mut Context<'input>,
 ) -> Result<()> {
-    #[cfg(not(feature = "positions"))]
-    let _ = token_span;
-    #[cfg(feature = "positions")]
-    let pos = token_span.start();
-
     let value = normalize_attribute(value, ctx)?;
 
-    if prefix.as_str() == XMLNS {
+    if prefix == XMLNS {
         // The xmlns namespace MUST NOT be declared as the default namespace.
         if value.as_str() == NS_XMLNS_URI {
-            let pos = ctx.err_pos_from_qname(prefix, local);
+            let pos = ctx.err_pos_at(attr_pos);
             return Err(Error::UnexpectedXmlnsUri(pos));
         }
 
@@ -649,15 +679,15 @@ fn process_attribute<'input>(
         // The prefix 'xml' is by definition bound to the namespace name
         // http://www.w3.org/XML/1998/namespace.
         // It MUST NOT be bound to any other namespace name.
-        if local.as_str() == NS_XML_PREFIX {
+        if local == NS_XML_PREFIX {
             if !is_xml_ns_uri {
-                let pos = ctx.err_pos_from_span(prefix);
+                let pos = ctx.err_pos_at(attr_pos);
                 return Err(Error::InvalidXmlPrefixUri(pos));
             }
         } else {
             // The xml namespace MUST NOT be bound to a non-xml prefix.
             if is_xml_ns_uri {
-                let pos = ctx.err_pos_from_span(prefix);
+                let pos = ctx.err_pos_at(attr_pos);
                 return Err(Error::UnexpectedXmlUri(pos));
             }
         }
@@ -666,26 +696,26 @@ fn process_attribute<'input>(
         if ctx
             .doc
             .namespaces
-            .exists(ctx.namespace_start_idx, Some(local.as_str()))
+            .exists(ctx.namespace_start_idx, Some(local))
         {
-            let pos = ctx.err_pos_from_qname(prefix, local);
-            return Err(Error::DuplicatedNamespace(local.as_str().to_string(), pos));
+            let pos = ctx.err_pos_at(attr_pos);
+            return Err(Error::DuplicatedNamespace(local.to_string(), pos));
         }
 
         // Xml namespace should not be added to the namespaces.
         if !is_xml_ns_uri {
-            ctx.doc.namespaces.push_ns(Some(local.as_str()), value)?;
+            ctx.doc.namespaces.push_ns(Some(local), value)?;
         }
-    } else if local.as_str() == XMLNS {
+    } else if local == XMLNS {
         // The xml namespace MUST NOT be declared as the default namespace.
         if value.as_str() == NS_XML_URI {
-            let pos = ctx.err_pos_from_span(local);
+            let pos = ctx.err_pos_at(attr_pos);
             return Err(Error::UnexpectedXmlUri(pos));
         }
 
         // The xmlns namespace MUST NOT be declared as the default namespace.
         if value.as_str() == NS_XMLNS_URI {
-            let pos = ctx.err_pos_from_span(local);
+            let pos = ctx.err_pos_at(attr_pos);
             return Err(Error::UnexpectedXmlnsUri(pos));
         }
 
@@ -696,7 +726,7 @@ fn process_attribute<'input>(
             local,
             value,
             #[cfg(feature = "positions")]
-            pos,
+            pos: attr_pos,
         });
     }
 
@@ -704,8 +734,8 @@ fn process_attribute<'input>(
 }
 
 fn process_element<'input>(
-    end_token: xmlparser::ElementEnd<'input>,
-    token_span: StrSpan<'input>,
+    end_token: tokenizer::ElementEnd<'input>,
+    token_range: Range<usize>,
     ctx: &mut Context<'input>,
 ) -> Result<()> {
     if ctx.tag_name.name.is_empty() {
@@ -713,12 +743,12 @@ fn process_element<'input>(
         // <!DOCTYPE test [ <!ENTITY p '</p>'> ]>
         // <root>&p;</root>
 
-        if let xmlparser::ElementEnd::Close(..) = end_token {
+        if let tokenizer::ElementEnd::Close(..) = end_token {
             return Err(Error::UnexpectedEntityCloseTag(
-                ctx.err_pos_from_span(token_span),
+                ctx.err_pos_at(token_range.start),
             ));
         } else {
-            unreachable!("should be already checked by the xmlparser");
+            unreachable!("should be already checked by the tokenizer");
         }
     }
 
@@ -728,25 +758,27 @@ fn process_element<'input>(
     let attributes = resolve_attributes(namespaces, ctx)?;
 
     match end_token {
-        xmlparser::ElementEnd::Empty => {
-            let tag_ns_idx = get_ns_idx_by_prefix(namespaces, ctx.tag_name.prefix, ctx)?;
+        tokenizer::ElementEnd::Empty => {
+            let tag_ns_idx = get_ns_idx_by_prefix(
+                namespaces,
+                ctx.tag_name.prefix_pos,
+                ctx.tag_name.prefix,
+                ctx,
+            )?;
             let new_element_id = ctx.append_node(
                 NodeKind::Element {
                     tag_name: ExpandedNameIndexed {
                         namespace_idx: tag_ns_idx,
-                        local_name: ctx.tag_name.name.as_str(),
+                        local_name: ctx.tag_name.name,
                     },
                     attributes,
                     namespaces,
                 },
-                ctx.tag_name.span.start()..token_span.end(),
+                ctx.tag_name.pos..token_range.end,
             )?;
             ctx.awaiting_subtree.push(new_element_id);
         }
-        xmlparser::ElementEnd::Close(prefix, local) => {
-            let prefix = prefix.as_str();
-            let local = local.as_str();
-
+        tokenizer::ElementEnd::Close(prefix, local) => {
             let parent_node = &mut ctx.doc.nodes[ctx.parent_id.get_usize()];
             // should never panic as we start with the single prefix of the
             // root node and always push another one when changing the parent
@@ -754,16 +786,16 @@ fn process_element<'input>(
 
             #[cfg(feature = "positions")]
             {
-                parent_node.range.end = token_span.end();
+                parent_node.range.end = token_range.end;
             }
 
             if let NodeKind::Element { ref tag_name, .. } = parent_node.kind {
                 if prefix != parent_prefix || local != tag_name.local_name {
-                    return Err(Error::UnexpectedCloseTag {
-                        expected: gen_qname_string(parent_prefix, tag_name.local_name),
-                        actual: gen_qname_string(prefix, local),
-                        pos: ctx.err_pos_from_span(token_span),
-                    });
+                    return Err(Error::UnexpectedCloseTag(
+                        gen_qname_string(parent_prefix, tag_name.local_name),
+                        gen_qname_string(prefix, local),
+                        ctx.err_pos_at(token_range.start),
+                    ));
                 }
             }
             ctx.awaiting_subtree.push(ctx.parent_id);
@@ -773,23 +805,28 @@ fn process_element<'input>(
                 ctx.parent_prefixes.pop();
                 debug_assert!(!ctx.parent_prefixes.is_empty());
             } else {
-                unreachable!("should be already checked by the xmlparser");
+                unreachable!("should be already checked by the tokenizer");
             }
         }
-        xmlparser::ElementEnd::Open => {
-            let tag_ns_idx = get_ns_idx_by_prefix(namespaces, ctx.tag_name.prefix, ctx)?;
+        tokenizer::ElementEnd::Open => {
+            let tag_ns_idx = get_ns_idx_by_prefix(
+                namespaces,
+                ctx.tag_name.prefix_pos,
+                ctx.tag_name.prefix,
+                ctx,
+            )?;
             ctx.parent_id = ctx.append_node(
                 NodeKind::Element {
                     tag_name: ExpandedNameIndexed {
                         namespace_idx: tag_ns_idx,
-                        local_name: ctx.tag_name.name.as_str(),
+                        local_name: ctx.tag_name.name,
                     },
                     attributes,
                     namespaces,
                 },
-                ctx.tag_name.span.start()..token_span.end(),
+                ctx.tag_name.pos..token_range.end,
             )?;
-            ctx.parent_prefixes.push(ctx.tag_name.prefix.as_str());
+            ctx.parent_prefixes.push(ctx.tag_name.prefix);
         }
     }
 
@@ -823,10 +860,7 @@ impl Context<'_> {
     }
 }
 
-fn resolve_attributes<'input>(
-    namespaces: ShortRange,
-    ctx: &mut Context<'input>,
-) -> Result<ShortRange> {
+fn resolve_attributes(namespaces: ShortRange, ctx: &mut Context) -> Result<ShortRange> {
     if ctx.current_attributes.is_empty() {
         return Ok(ShortRange::new(0, 0));
     }
@@ -839,7 +873,7 @@ fn resolve_attributes<'input>(
 
     let current_attributes = core::mem::take(&mut ctx.current_attributes);
     for attr in current_attributes {
-        let namespace_idx = if attr.prefix.as_str() == NS_XML_PREFIX {
+        let namespace_idx = if attr.prefix == NS_XML_PREFIX {
             // The prefix 'xml' is by definition bound to the namespace name
             // http://www.w3.org/XML/1998/namespace. This namespace is added
             // to the document on creation and is always element 0.
@@ -849,19 +883,19 @@ fn resolve_attributes<'input>(
             // always has no value.'
             None
         } else {
-            get_ns_idx_by_prefix(namespaces, attr.prefix, ctx)?
+            get_ns_idx_by_prefix(namespaces, attr.pos, attr.prefix, ctx)?
         };
 
         let attr_name = ExpandedNameIndexed {
             namespace_idx,
-            local_name: attr.local.as_str(),
+            local_name: attr.local,
         };
 
         // Check for duplicated attributes.
         if ctx.doc.attributes[start_idx..].iter().any(|attr| {
             attr.name.as_expanded_name(&ctx.doc) == attr_name.as_expanded_name(&ctx.doc)
         }) {
-            let pos = ctx.err_pos_from_qname(attr.prefix, attr.local);
+            let pos = ctx.err_pos_at(attr.pos);
             return Err(Error::DuplicatedAttribute(attr.local.to_string(), pos));
         }
 
@@ -876,17 +910,21 @@ fn resolve_attributes<'input>(
     Ok((start_idx..ctx.doc.attributes.len()).into())
 }
 
-fn process_text<'input>(text: StrSpan<'input>, ctx: &mut Context<'input>) -> Result<()> {
+fn process_text<'input>(
+    text: &'input str,
+    range: Range<usize>,
+    ctx: &mut Context<'input>,
+) -> Result<()> {
     // Add text as is if it has only valid characters.
-    if !text.as_str().bytes().any(|b| b == b'&' || b == b'\r') {
-        append_text(StringStorage::Borrowed(text.as_str()), text.range(), ctx)?;
+    if !text.bytes().any(|b| b == b'&' || b == b'\r') {
+        append_text(StringStorage::Borrowed(text), range, ctx)?;
         ctx.after_text = true;
         return Ok(());
     }
 
     let mut text_buffer = String::with_capacity(32);
     let mut is_as_is = false; // TODO: explain
-    let mut stream = Stream::from_substr(ctx.doc.text, text.range());
+    let mut stream = Stream::from_substr(ctx.doc.text, range.clone());
     while !stream.at_end() {
         match parse_next_chunk(&mut stream, &ctx.entities)? {
             NextChunk::Byte(c) => {
@@ -914,7 +952,7 @@ fn process_text<'input>(text: StrSpan<'input>, ctx: &mut Context<'input>) -> Res
 
                 if !text_buffer.is_empty() {
                     let storage = StringStorage::new_owned(text_buffer.as_str());
-                    append_text(storage, text.range(), ctx)?;
+                    append_text(storage, range.clone(), ctx)?;
                     text_buffer.clear();
                     ctx.after_text = true;
                 }
@@ -922,12 +960,12 @@ fn process_text<'input>(text: StrSpan<'input>, ctx: &mut Context<'input>) -> Res
                 ctx.loop_detector.inc_references(&stream)?;
                 ctx.loop_detector.inc_depth(&stream)?;
 
-                let parser = xmlparser::Tokenizer::from_fragment(ctx.doc.text, fragment.range());
+                let mut stream = Stream::from_substr(ctx.doc.text, fragment.range());
                 let prev_tag_name = ctx.tag_name;
                 ctx.tag_name = TagNameSpan::new_null();
-                process_tokens(parser, ctx)?;
-                text_buffer.clear();
+                tokenizer::parse_content(&mut stream, ctx)?;
                 ctx.tag_name = prev_tag_name;
+                text_buffer.clear();
 
                 ctx.loop_detector.dec_depth();
             }
@@ -935,7 +973,7 @@ fn process_text<'input>(text: StrSpan<'input>, ctx: &mut Context<'input>) -> Res
     }
 
     if !text_buffer.is_empty() {
-        append_text(StringStorage::new_owned(text_buffer), text.range(), ctx)?;
+        append_text(StringStorage::new_owned(text_buffer), range, ctx)?;
         ctx.after_text = true;
     }
 
@@ -945,27 +983,27 @@ fn process_text<'input>(text: StrSpan<'input>, ctx: &mut Context<'input>) -> Res
 // While the whole purpose of CDATA is to indicate to an XML library that this text
 // has to be stored as is, carriage return (`\r`) is still has to be replaced with `\n`.
 fn process_cdata<'input>(
-    text: StrSpan<'input>,
-    span: StrSpan<'input>,
+    text: &'input str,
+    range: Range<usize>,
     ctx: &mut Context<'input>,
 ) -> Result<()> {
     // Add text as is if it has only valid characters.
-    if !text.as_str().as_bytes().contains(&b'\r') {
-        append_text(StringStorage::Borrowed(text.as_str()), span.range(), ctx)?;
+    if !text.as_bytes().contains(&b'\r') {
+        append_text(StringStorage::Borrowed(text), range, ctx)?;
         ctx.after_text = true;
         return Ok(());
     }
 
     let mut text_buffer = String::with_capacity(32);
-    let count = text.as_str().chars().count();
-    for (i, c) in text.as_str().chars().enumerate() {
+    let count = text.chars().count();
+    for (i, c) in text.chars().enumerate() {
         for b in CharToBytes::new(c) {
             text_buffer.push_from_text(b, i + 1 == count);
         }
     }
 
     if !text_buffer.is_empty() {
-        append_text(StringStorage::new_owned(text_buffer), text.range(), ctx)?;
+        append_text(StringStorage::new_owned(text_buffer), range, ctx)?;
         ctx.after_text = true;
     }
 
@@ -1117,8 +1155,9 @@ fn _normalize_attribute(text: StrSpan, buffer: &mut String, ctx: &mut Context) -
 }
 
 fn get_ns_idx_by_prefix(
-    range: ShortRange,
-    prefix: StrSpan,
+    namespaces: ShortRange,
+    prefix_pos: usize,
+    prefix: &str,
     ctx: &Context,
 ) -> Result<Option<NamespaceIdx>> {
     // Prefix CAN be empty when the default namespace was defined.
@@ -1128,10 +1167,10 @@ fn get_ns_idx_by_prefix(
     let prefix_opt = if prefix.is_empty() {
         None
     } else {
-        Some(prefix.as_str())
+        Some(prefix)
     };
 
-    let idx = ctx.doc.namespaces.tree_order[range.to_urange()]
+    let idx = ctx.doc.namespaces.tree_order[namespaces.to_urange()]
         .iter()
         .find(|idx| ctx.doc.namespaces.get(**idx).name == prefix_opt);
 
@@ -1144,8 +1183,8 @@ fn get_ns_idx_by_prefix(
                 //
                 // Example:
                 // <e random:a='b'/>
-                let pos = ctx.err_pos_from_span(prefix);
-                Err(Error::UnknownNamespace(prefix.as_str().to_string(), pos))
+                let pos = ctx.err_pos_at(prefix_pos);
+                Err(Error::UnknownNamespace(prefix.to_string(), pos))
             } else {
                 // If an URI was not found and prefix IS empty than
                 // an element or an attribute doesn't have a namespace.