+++ testing-tools/afl-fuzz/src/main.rs
@@ -1,13 +1,6 @@
-extern crate afl;
-extern crate roxmltree;
-
-use std::str;
-
-use afl::fuzz;
-
 fn main() {
-    fuzz(|data| {
-        if let Ok(text) = str::from_utf8(data) {
+    afl::fuzz!(|data: &[u8]| {
+        if let Ok(text) = std::str::from_utf8(data) {
             let _ = roxmltree::Document::parse(&text);
         }
     });